refacto: use dbPsychologists to get declared university
